fixed doctest

+++ src/autora/experiment_runner/synthetic/abstract/template_experiment.py
@@ -28,7 +28,7 @@ Examples:
     functools.partial(<function template_experiment.<locals>.run at 0x...>,
                       added_noise=0.0)
 
-    >>> s.ground_truth(1.)
+    >>> np.float64(s.ground_truth(1.))
     np.float64(2.0)
 
     >>> s.ground_truth(s.domain())
@@ -41,7 +41,7 @@ Examples:
     >>> s.run  # doctest: +ELLIPSIS +NORMALIZE_WHITESPACE
     <function template_experiment.<locals>.run at 0x...>
 
-    >>> s.run(1., random_state=42)
+    >>> np.float64(s.run(1., random_state=42))
     np.float64(2.003047170797544)
 
     >>> s.run(s.domain(), random_state=42)
